+++ docs/changelog.md
@@ -1,4 +1,9 @@
 # TBD
+### Features
+* Added wait-for-availability checks to the indexer, wallet, contract-helper service, explorer-wamp, and explorer
+
+### Fixes
+* Fixed an issue with the indexer database not getting created
 
 # 0.5.18
 ### Features